# docs/swagger.yaml
definitions:
...
  server.ApiResponse:
    properties:
      code:
        description: 状态码
        type: integer
      message:
        description: 状态短语
        type: string
      result:
        description: 数据结果集
        type: object
    type: object
...
  server.IdStruct:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  server.OnlineVersionVO:
    properties:
      app_hash:
        type: string
      app_id:
        description: 应用id
        type: integer
      created_at:
        type: string
      domain:
        description: 应用域名
        type: string
      id:
        type: integer
      note:
        description: 版本介绍
        type: string
      status:
        description: 状态值
        type: integer
      updated_at:
        type: string
      version:
        description: 版本号
        type: string
    type: object
...
  server.VersionArg:
    properties:
      description:
        type: string
      domain:
        type: string
      random:
        type: string
      version:
        type: string
    type: object
  server.VersionListVO:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/version.Version'
        type: array
    required:
    - count
    - list
    type: object
  server.VersionVO:
    properties:
      domain:
        type: string
      hash:
        type: string
    type: object
...
  version.Version:
    properties:
      app_id:
        description: 应用id
        type: integer
      created_at:
        type: string
      domain:
        description: 应用域名
        type: string
      id:
        type: integer
      note:
        description: 版本介绍
        type: string
      status:
        description: 状态值
        type: integer
      updated_at:
        type: string
      version:
        description: 版本号
        type: string
    type: object
info:
  contact: {}
paths:
...
  /version/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: string
      - description: 页容量
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  $ref: '#/definitions/server.VersionListVO'
              type: object
      summary: 版本列表
      tags:
      - version
  /version/offline:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 请求体
        in: body
        name: IdStruct
        required: true
        schema:
          $ref: '#/definitions/server.IdStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  type: object
              type: object
      summary: 版本下线
      tags:
      - version
  /version/online:
    get:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  $ref: '#/definitions/server.OnlineVersionVO'
              type: object
      summary: 上线版本
      tags:
      - version
  /version/publish:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 请求体
        in: body
        name: VersionArg
        required: true
        schema:
          $ref: '#/definitions/server.VersionArg'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  $ref: '#/definitions/server.VersionVO'
              type: object
      summary: 发布新版本
      tags:
      - version
swagger: "2.0"